+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to QETLAB will be documented in this file.
 - BCSGameLB: Computes a lower bound on the quantum value of a binary constraint system (BCS) game.
 - BCSGameValue: Computes the maximum value of a binary constraint system (BCS) game. In classical and no-signalling settings, the value computed is exact, but the quantum value is just an upper bound.
 - InducedMatrixNorm: Computes a lower bound on the induced p->q norm of a matrix.
+- InducedSchattenNorm: Computes a lower bound on the induced Schatten p->q norm of a superoperator.
 - NonlocalGameLB: Computes a lower bound on the quantum value of a two-player non-local game.
 - RandomPOVM: Computes a random POVM of a specified size and with a specified number of outcomes.
 - helpers/bcs_to_nonlocal: Converts a description of a binary constraint system (BCS) game into a form that can be presented as a general non-local game.